requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-dateutil
google-api-python-client
